requirements.txt
Flask
faiss-cpu
sentence-transformers
python-docx
PyMuPDF
PyMuPDF
docx
sentence_transformers